[vm] Hide local variables synthesized in the front-end from debugger
TEST=pkg/vm_service/test/patterns_local_vars_test.dart

Fixes https://github.com/dart-lang/sdk/issues/51512

Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/288262
Commit-Queue: Alexander Markov <alexmarkov@google.com>

+++ runtime/vm/compiler/frontend/kernel_translation_helper.h
@@ -447,7 +447,7 @@ class VariableDeclarationHelper {
 
   TokenPosition position_ = TokenPosition::kNoSource;
   TokenPosition equals_position_ = TokenPosition::kNoSource;
-  uint8_t flags_ = 0;
+  uint32_t flags_ = 0;
   StringIndex name_index_;
   intptr_t annotation_count_ = 0;
